Chore: Revert mobile-specific logic
