+++ packages/app-mobile/components/screens/Note/commands/index.ts
@@ -3,7 +3,6 @@ import * as attachFile from './attachFile';
 import * as hideKeyboard from './hideKeyboard';
 import * as insertDateTime from './insertDateTime';
 import * as setTags from './setTags';
-import * as shareCurrentNote from './shareCurrentNote';
 import * as toggleVisiblePanes from './toggleVisiblePanes';
 
 const index: any[] = [
@@ -11,7 +10,6 @@ const index: any[] = [
 	hideKeyboard,
 	insertDateTime,
 	setTags,
-	shareCurrentNote,
 	toggleVisiblePanes,
 ];
