+++ packages/app-mobile/root.tsx
@@ -522,7 +522,6 @@ async function initialize(dispatch: Dispatch) {
 	Setting.setConstant('cacheDir', `${getProfilesRootDir()}/cache`);
 	const resourceDir = getResourceDir(currentProfile, isSubProfile);
 	Setting.setConstant('resourceDir', resourceDir);
-	Setting.setConstant('pluginAssetDir', `${Setting.value('resourceDir')}/pluginAssets`);
 	Setting.setConstant('pluginDir', `${getProfilesRootDir()}/plugins`);
 	Setting.setConstant('pluginDataDir', getPluginDataDir(currentProfile, isSubProfile));
